requirements.txt
alembic==1.18.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
noise==1.2.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
setuptools==80.9.0
SQLAlchemy==2.0.46
termcolor==3.3.0
typing_extensions==4.15.0
Werkzeug==3.1.3
wheel==0.45.1
